Move AgencyProfileForm to components folder
