+++ frontend/src/components/AgencyProfileForm/AgencyProfileForm.js
@@ -2,18 +2,18 @@ import React, { Component } from "react";
 import { withRouter } from "react-router-dom";
 import FormSectionHeader from "./FormSectionHeader";
 import { FormRow, FormCol } from "./FormLayout";
-import InputText from "../../components/FormComponents/InputText";
-import InputDate from "../../components/FormComponents/InputDate";
-import InputDropdown from "../../components/FormComponents/InputDropdown";
+import InputText from "../FormComponents/InputText";
+import InputDate from "../FormComponents/InputDate";
+import InputDropdown from "../FormComponents/InputDropdown";
 import CheckboxList from "./CheckboxList";
 import IncrementerBoxList from "./IncrementerBoxList";
-import FormButton from "../../components/FormComponents/FormButton";
+import FormButton from "../FormComponents/FormButton";
 import SmallButton from "./SmallButton";
 import AddressList from "./AddressList";
 import ContactsList from "./ContactsList";
 import DistributionDays from "./DistributionDays";
 import Calendar from "./DistributionCalendar/Calendar";
-import InlineDropdown from "../../components/FormComponents/InlineDropdown";
+import InlineDropdown from "../FormComponents/InlineDropdown";
 import "typeface-roboto";
 import "./FormStyle.css";
 import { getJWT } from "../../auth";
