+++ frontend/src/components/FormComponents/TimeBoxes.css
@@ -1,0 +1,25 @@
+.time-input {
+  display: block;
+  position: relative;
+  width: auto;
+  float: right;
+  margin-top: 9px;
+}
+
+.number-field {
+  position: relative;
+  height: 25px;
+  width: 35px;
+  display: inline-block;
+  padding: 0px;
+  padding-right: 3px;
+  margin: 0px 2px 0px 2px;
+  border-radius: 2px;
+  border: 1px solid #000000;
+  font-size: 15px;
+  text-align: right;
+}
+
+.number-field::placeholder {
+  color: #cdcbd4;
+}
